apps/api_auth/views.py
from rest_framework.response import Response
from rest_framework.views import APIView
from rest_framework_simplejwt.tokens import RefreshToken
from django.contrib.auth import authenticate
from .serializers import LoginSerializer
from rest_framework.exceptions import ValidationError


class LoginView(APIView):
    def post(self, request):
        LoginSerializer(data=request.data).is_valid(raise_exception=True)
        username = request.data.get("username")
        password = request.data.get("password")
        user = authenticate(request, username=username, password=password)

        if user is None:
            raise ValidationError({"password": ["Invalid credentials"]})

        refresh = RefreshToken.for_user(user)
        return Response(
            {
                "refresh": str(refresh),
                "access": str(refresh.access_token),
            }
        )
