requirements.txt
asgiref==3.8.1
Django==5.1.2
django-environ==0.11.2
moomoo-api==8.5.4508
numpy==2.0.1
pandas==2.2.2
protobuf==3.20.3
pycryptodome==3.20.0
python-dateutil==2.9.0.post0
pytz==2024.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
psycopg2-binary>=2.9.1
djangorestframework
djangorestframework-simplejwt
django-cors-headers
